requirements.txt
PyYAML>=6.0
PyQt6>=6.0; python_version>='3.9' 